renderers: move type_renderers into own package
madore should not depend on all packages for which it provides
renderers. People should not have to install matplotlib unless they want
to draw a plot.

Change the previous approach of wrapping every type renderer in a
try/except block. This was hard to read and ignored all ImportErrors
indiscriminately.

Instead, hook into python's import mechanism and load renderers only
when the corresponding package is loaded. They can't be used before
anyway. This means that renderers need to be named after the packages
now.